requirements.txt
netifaces
click
python-nmap
scapy
